Changed grouping so that now only those SNPs that share either a gene or a pdbid are grouped for combination. Now logging plink frequency output like LD and Fst

+++ lib/snp_stats.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python27
 
 # Command Line Arguments:
-# calc_fst.r pop-id-map pedmap krange snp-groups
+# calc_fst.r pop-id-map pedmap krange snp-pdbid-loc
 
 import sys,os,csv,subprocess,math
 from itertools import combinations,product
@@ -10,7 +10,9 @@ def main():
 
   ## Check cmd line arguments
   if len(sys.argv) < 2:
-    print "usage: \n\tcalc_fst.py pop-id-map pedmap krange snp-groups snp-pdb-loc"
+    print "usage: \n\tcalc_fst.py pop-id-map pedmap krange snp-pdbid-loc"
+    print "\tpop-id-map is a tab delimited file with [population, sampleid] columns"
+    print "\tsnp-pdbid-loc is a tab delimited file with [gene, snp_name, pdbid, chain, x, y , z] columns"
     sys.exit(1)
 
   ## Read cmd line arguments
@@ -36,48 +38,80 @@ def main():
 
   ## Remove SNPs with MAF == 0.0
   snp_map = qc_freq(snp_map,pedmap)
+  print "\n# %d SNPs found in PEDMAP."%len(snp_map)
 
   ## Read SNP groups if specified
   if snp_groups:
     group_snp_dict,snp_loc,snp_list = read_groups(snp_groups,snp_map)
+    print "# Evaluating %d overlapping SNPs."%len(snp_list)
   # Only group-associated SNPs present in the snp_map are recorded
   # As such, snp_list holds the union of the pedmap and pdbmap SNPs
 
   ## Compute pairwise distance and LD (R^2) for all SNPs
+  print "\n# Calculating LD statistics and 1D distances..."
   snp_ld,snp_bp_dist = calc_ld_dist(pedmap,snp_list)
 
   ## Compute pairwise structural distance for all SNPs
+  print "# Calculating 3D distances..."
   snp_3d_dist = calc_3d_dist(snp_loc)
 
   ## Compute the global and individual Fst scores, initialize snp_fstat
+  print "# Calculating global Fst estimates..."
   snp_fstat = calc_global_fst(snp_list,pop_id_map,pedmap)
   if 1 in krange: krange.remove(1) # k=1 implicitly calculated
 
   ## Compute Fst for k-wise tuples (k > 1), update snp_fstat
-  # snp_fstat = calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_groups,group_snp_dict)
-  snp_fstat = calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_list)
+  print "# Calculating k-wise Fst estimates..."
+  snp_fstat = calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_groups,group_snp_dict)
+  # snp_fstat = calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_list)
 
   ## Aggregate all the measurements and statistics
-  aggregate_stats(snp_bp_dist,snp_3d_dist,snp_ld,snp_fstat)
+  print "# Aggregating results..."
+  res = aggregate_stats(snp_bp_dist,snp_3d_dist,snp_ld,snp_fstat)
+
+  ## Save all results
+  print "# Saving results..."
+  save_results(pedmap,res)
+
+def save_results(pedmap,res):
+  # create results directory
+  resdir = 'results/%s'%pedmap
+  if not os.path.exists(resdir):
+    os.makedirs(resdir)
+  # copy the scripts used
+  os.system("cp -f lib/snp_stats.py %s/"%resdir)
+  os.system("cp -f `which calc_fst.r` %s/"%resdir)
+  # record the original command
+  with open("%s/command.txt"%resdir,'wb') as fout:
+    fout.write("\n".join(sys.argv))
+  # save the results
+  with open("%s/%s.stats"%(resdir,pedmap)) as fout:
+    writer = csv.writer(fout,delimiter='\t')
+    for row in res:
+      writer.writerow(row)
 
 def aggregate_stats(snp_1d_dist,snp_3d_dist,snp_ld,snp_fst):
   # all snps with 3d structural information and a partner
+  res = []
   snps = snp_3d_dist.keys()
-  print "rs# Fst(ind) Fst(n(1d)) Fst(n(3d)) LD(n(1d)) LD(n(3d))"
+  print "rsID\tFst(ind)\tn(1d)\tn(3d)\tFst(n(1d))\tFst(n(3d))\tLD(n(1d))\tLD(n(3d))"
   for snp in snps:
     ind_fst = snp_fst[snp][1]
     # get the snp name of the nearest snp in 1d and 3d
-    nearest1d = min(snp_1d_dist[snp],key=snp_1d_dist[snp].get)
+    n1d = min(snp_1d_dist[snp],key=snp_1d_dist[snp].get)
     # for all partner snps B, find shortest distance b/w A and B
     # and select the A->B pair with the shortest distance overall
-    nearest3d = min([(min(snp_3d_dist[snp][snpb][0]),snpb) for snpb in snp_3d_dist[snp]])[1]
+    n3d = min([(min(snp_3d_dist[snp][snpb][0]),snpb) for snpb in snp_3d_dist[snp]])[1]
     # Lookup Fst for nearest partners
-    fstn1d = snp_fst[snp][2][(nearest1d,)]
-    fstn3d = float("NaN") if nearest3d == "NA" else snp_fst[snp][2][(nearest3d,)]
+    fstn1d = snp_fst[snp][2][(n1d,)]
+    fstn3d = float("NaN") if n3d == "NA" else snp_fst[snp][2][(n3d,)]
     # Lookup LD for nearest partners
-    ldn1d  = snp_ld[snp][nearest1d]
-    ldn3d  = float("NaN") if nearest3d == "NA" else snp_ld[snp][nearest3d]
-    print "%s %f %f %f %f %f"%(snp,ind_fst,fstn1d,fstn3d,ldn1d,ldn3d)
+    ldn1d  = snp_ld[snp][n1d]
+    ldn3d  = float("NaN") if n3d == "NA" else snp_ld[snp][n3d]
+    stats  = [snp,ind_fst,n1d,n3d,fstn1d,fstn3d,ldn1d,ldn3d]
+    print '\t'.join(stats)
+    res.append(stats)
+  return res
 
 def read_pop(pop_id_map):
   with open(pop_id_map,'rb') as fin:
@@ -87,7 +121,7 @@ def read_pop(pop_id_map):
     for row in reader:
       pop_id_dict.setdefault(row[1],[]).append(row[0])
   populations = pop_id_dict.keys()
-  print "\nPopulations: %s"%" ".join([str((i,pop)) for i,pop in enumerate(populations)])
+  print "\n# Populations: %s"%" ".join([str((i,pop)) for i,pop in enumerate(populations)])
   return pop_id_dict
 
 def read_map(pedmap):
@@ -101,22 +135,26 @@ def read_map(pedmap):
   return snp_map
 
 def qc_freq(snp_map,pedmap):
-  try:
-    cmd = "plink --file %s --freq --noweb"%pedmap
-    p = subprocess.Popen(cmd,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-    out,err = p.communicate() # ensure process has finished
-    with open("plink.frq",'rb') as fin:
-      header = fin.readline()
-      for line in fin:
-        row = line.split()
-        snp,frq = row[1],float(row[4])
-        if frq < 0.0000001: # "zero"
-          del snp_map[snp]
+  # If LD has been calculated before, load from log, else compute
+  if not os.path.isfile("plink.frq %s.frq"%pedmap):
+    try:
+      cmd = "plink --file %s --freq --noweb"%pedmap
+      p = subprocess.Popen(cmd,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+      out,err = p.communicate() # ensure process has finished
+      os.system("cp -f plink.frq %s.frq"%pedmap)
+    except: raise
+    finally:
+      os.system("rm -f plink.*") # remove plink output files
+  print "# Loading frequency from log..."
+  with open("%s.frq"%pedmap,'rb') as fin:
+    header = fin.readline()
+    for line in fin:
+      row = line.split()
+      snp,frq = row[1],float(row[4])
+      if frq < 0.0000001: # "zero"
+        del snp_map[snp]
     return snp_map
-  except: raise
-  finally:
-    os.system("rm -f plink.*") # remove plink output files
-
+  
 def read_groups(snp_groups,snp_map):
   snp_loc = {}
   all_snps = set([])
@@ -125,21 +163,23 @@ def read_groups(snp_groups,snp_map):
     group_snp_dict = {}
     # pdbid : [snp,...]
     with open(snp_groups,'rb') as fin:
+      fin.readline() # burn the header
+      # gene,snp_name,pdbid,chain,x,y,z
       reader = csv.reader(fin,delimiter='\t')
       for row in reader:
-        all_snps.add(row[0])
+        all_snps.add(row[1])
         # Only read SNPs present in the pedmap
-        if row[0] in snp_map:
+        if row[1] in snp_map:
           # Only read unique SNPs (intentional duplication in file)
-          if row[0] not in group_snp_dict.setdefault(row[1],[]):
-            group_snp_dict[row[1]].append(row[0])
+          if row[1] not in group_snp_dict.setdefault(row[2],[]):
+            group_snp_dict[row[2]].append(row[1]) # store the pdbid-snp association
+          if row[1] not in group_snp_dict.setdefault(row[0],[]):
+            group_snp_dict[row[0]].append(row[1]) # store the gene-snp association
           # Record the SNP (qualified by pdbid and chain) 3D location
-          snp_loc.setdefault(row[0],{})[(row[1],row[2])] = [float(x) for x in row[3:]]
-          union_snps.add(row[0])
+          snp_loc.setdefault(row[1],{})[(row[2],row[3])] = [float(x) for x in row[4:]]
+          union_snps.add(row[1])
     groups = group_snp_dict.keys()
-    print "\nGroups: %s"%" ".join(groups)
-    print "\n%d of %d group SNPs found in pedmap with MAF > 0.0."%(
-              len(union_snps),len(all_snps))
+  print "# %d SNPs found in PDBMap."%len(all_snps)
   # Pull those SNPs 
   snp_list = list(union_snps)
   snp_list.sort()
@@ -148,39 +188,43 @@ def read_groups(snp_groups,snp_map):
 def calc_ld_dist(pedmap,snp_list):
   snp_bp_dist  = {}
   snp_ld = {}
-  try:
-    # cmd = "plink --file %s --r2 --ld-window-kb 500000 --noweb"%pedmap
-    cmd = "plink --file %s --r2 --ld-window 250000000 --ld-window-kb 250000 --ld-window-r2 0.0 --noweb"%pedmap
-    p = subprocess.Popen(cmd,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-    out,err = p.communicate() # ensure process has finished
-    with open("plink.ld",'rb') as fin:
-      header = fin.readline()
-      # CHR_A BP_A SNP_A CHR_B BP_B SNP_B R2
-      for line in fin:
-        row = line.split()
-        # Only include pedmap-pdbmap union SNPs
-        # and if the SNPs are on the same chromosome
-        if row[2] in snp_list and row[5] in snp_list and row[0]==row[3]:
-          row[1],row[4],row[6] = float(row[1]),float(row[4]),float(row[6])
-          # Distances
-          snp_bp_dist.setdefault(row[2],{})[row[5]] = math.fabs(row[4]-row[1])
-          snp_bp_dist.setdefault(row[5],{})[row[2]] = math.fabs(row[4]-row[1])
-          # LD (R^2)
-          snp_ld.setdefault(row[2],{})[row[5]] = row[6]
-          snp_ld.setdefault(row[5],{})[row[2]] = row[6]
 
-    snpAs = snp_bp_dist.keys()
-    for snpA in snpAs:
-      snpBs = snp_bp_dist[snpA].keys()
-      # for snpB in snpBs:
-      #   print "%s %s -> GD=%d LD=%f"%(snpA,snpB,snp_bp_dist[snpA][snpB],snp_ld[snpA][snpB])
+  # If LD has been calculated before, load from log, else compute
+  if not os.path.isfile("plink.ld %s.ld"%pedmap):
+    try:
+      print "# Calculating LD with plink..."
+      cmd = "plink --file %s --r2 --ld-window 250000000 --ld-window-kb 250000 --ld-window-r2 0.0 --noweb"%pedmap
+      p = subprocess.Popen(cmd,shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+      out,err = p.communicate() # ensure process has finished
+      os.system("cp -f plink.ld %s.ld"%pedmap)
+    except: raise
+    finally:
+      os.system("rm -f plink.*") # remove plink output files
+  print "# Loading LD from log..."
+  with open("%s.ld"%pedmap,'rb') as fin:
+    header = fin.readline()
+    # CHR_A BP_A SNP_A CHR_B BP_B SNP_B R2
+    for line in fin:
+      row = line.split()
+      # Only include pedmap-pdbmap union SNPs
+      # and if the SNPs are on the same chromosome
+      if row[2] in snp_list and row[5] in snp_list and row[0]==row[3]:
+        row[1],row[4],row[6] = float(row[1]),float(row[4]),float(row[6])
+        # Distances
+        snp_bp_dist.setdefault(row[2],{})[row[5]] = math.fabs(row[4]-row[1])
+        snp_bp_dist.setdefault(row[5],{})[row[2]] = math.fabs(row[4]-row[1])
+        # LD (R^2)
+        snp_ld.setdefault(row[2],{})[row[5]] = row[6]
+        snp_ld.setdefault(row[5],{})[row[2]] = row[6]
 
-    print "# LD SNPs:",len(snp_ld)
-    print "# 1D SNPs:",len(snp_bp_dist)
-    return snp_ld,snp_bp_dist
-  except: raise
-  finally:
-    os.system("rm -f plink.*") # remove plink output files
+  snpAs = snp_bp_dist.keys()
+  for snpA in snpAs:
+    snpBs = snp_bp_dist[snpA].keys()
+
+  print "# # LD SNPs:",len(snp_ld)
+  print "# # 1D SNPs:",len(snp_bp_dist)
+  return snp_ld,snp_bp_dist
+    
 
 def calc_global_fst(snp_list,pop_id_map,pedmap):
   snp_fstat = {}
@@ -193,13 +237,12 @@ def calc_global_fst(snp_list,pop_id_map,pedmap):
     # Calculate Fst
     cmd = ["calc_fst.r",pop_id_map,pedmap,tempfile]
     p = subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-    out,err = p.communicate()
-    parser = string_parser(out.split('\n'))
+    parser = process_parser(p)
     global_fst,ind_fst,pop_freqs = parse_all(parser)
     # Collapse Fst and Frequencies into one dictionary
     snp_popgen = {}
-    print "# Fst scores:",len(ind_fst)
-    print "# Frequencies:",len(pop_freqs)
+    print "# # Fst scores:",len(ind_fst)
+    print "# # Frequencies:",len(pop_freqs)
     keys = pop_freqs.keys()
     keys.sort()
     for key in keys:
@@ -208,11 +251,9 @@ def calc_global_fst(snp_list,pop_id_map,pedmap):
     snps = snp_popgen.keys()
     snps.sort()
     for snp in snps:
-      # if not math.isnan(snp_popgen[snp][0]):
-      # print "%s\t%f"%(snp,snp_popgen[snp][0])
       snp_fstat[snp] = {1:snp_popgen[snp][0]}
-    print "# Fst SNPs:",len(snp_fstat)
-    print "Global Fst: ",global_fst
+    print "# # Fst SNPs:",len(snp_fstat)
+    print "# Global Fst: ",global_fst
     return snp_fstat # not necessary, but improves clarity
   except: raise
   finally:
@@ -225,7 +266,7 @@ def calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_list=None,snp_groups=N
 
     # If these have been computed before, load k-wise Fst from file
     if os.path.isfile('%s_%d-tuple.fst'%(pedmap,k)):
-      print "Loading %d-wise Fst from log..."%k
+      print "# Loading %d-wise Fst from log..."%k
       with open('%s_%d-tuple.fst'%(pedmap,k),'rb') as fin:
         reader = csv.reader(fin,delimiter='\t')
         for row in reader:
@@ -233,7 +274,7 @@ def calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_list=None,snp_groups=N
       continue # to next k
 
     # Generate k-tuples
-    print "Calculating tuples for k=%d..."%k
+    print "# Calculating tuples for k=%d..."%k
     combns = []
     if snp_groups:
       for group in group_snp_dict.iterkeys():
@@ -243,44 +284,43 @@ def calc_kwise_fst(snp_fstat,pop_id_map,pedmap,krange,snp_list=None,snp_groups=N
         combns = [x for x in combinations(snp_list,k)]
     else: raise Exception("Must provide snp_list or snp_groups")
 
+    # Some snp-snp duplication possible between groups, reduce to unique
+    unique_combns = [tup for tup in combns if tup[0] < tup[1]]
+    unqiue_combns.extend([tup for tup in combns if tup[1] < tup[0] and tup not in unique_combns])
+    del combns # immediately free memory
+
     # Write the k-tuples to temp file
     tempfile = "%d.temp"%multidigit_rand(5)
     try:
       with open(tempfile,'wb') as fout:
           writer = csv.writer(fout,delimiter=' ')
-          for combn in combns:
+          for combn in unqiue_combns:
             writer.writerow(combn)
 
       # Calculate Fst
       cmd = ["calc_fst.r",pop_id_map,pedmap,tempfile]
       p = subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-      out,err = p.communicate()
-      if err:
-        print "Errors:"
-        print err
-      parser = string_parser(out.split('\n'))
-      tuple_fstat[k] = parse_tuples(parser,combns)
+      parser = process_parser(p)
+      tuple_fstat[k] = parse_tuples(parser,unique_combns)
     except: raise
     finally:
       os.system("rm -f %s"%tempfile)
   
+  # Store all k-wise data into snp_fstat
   ks = tuple_fstat.keys()
   ks.sort()
   for k in ks:
     with open('%s_%d-tuple.fst'%(pedmap,k),'wb') as fout:
-      print "%d-tuples:"%k
+      print "# %d-tuples:"%k
       tups = tuple_fstat[k].keys()
       tups.sort()
       for tup in tups:
-        # if not math.isnan(tuple_fstat[k][tup][6]):
-        # print "%s\t%f"%('\t'.join(tup),tuple_fstat[k][tup][6])
         fst = tuple_fstat[k][tup][6]
         for snp in tup:
           partners = tuple([x for x in tup if x!=snp])
           if snp in snp_fstat:
             snp_fstat[snp].setdefault(k,{})[partners] = fst
             fout.write("%s\t%f\t%s\n"%(snp,fst,"\t".join(partners)))
-  print set([str(len([x for x in snp_fstat[snp][2]])) for snp in snp_fstat])
   return snp_fstat
 
 def calc_3d_dist(snp_loc):
@@ -302,13 +342,11 @@ def calc_3d_dist(snp_loc):
       snp_3d_dist[snp_pair[0]].setdefault(snp_pair[1],[])
       # Add distance, keep locations for reference
       snp_3d_dist[snp_pair[0]][snp_pair[1]].append((d,loc_pair))
-      # print "dist(%s,%s) = %2.3f via %s:%s->%s:%s"%(snp_pair[0],snp_pair[1],d,
-      #   loc_pair[0][0][0],loc_pair[0][0][1],loc_pair[1][0][0],loc_pair[1][0][1])
   lonely_snps = [snp for snp in snp_loc if snp not in snp_3d_dist]
   # Dummy code an NA partner and NaN distance for SNPs with no 3D partners
   for snp in lonely_snps:
     snp_3d_dist[snp] = {"NA":[(float("NaN"),None)]} 
-  print "# 3D SNPs:",len(snp_3d_dist)
+  print "# # 3D SNPs:",len(snp_3d_dist)
   return snp_3d_dist
 
 def parse_all(parser):
@@ -338,7 +376,7 @@ def parse_tuples(parser,combns):
     if len(line.split()) == 1:
       results[combns[combn_count]] = [float("NaN")]*11
     else:
-      results[combns[combn_count]] = [float(x) for x in line.split()]
+      results[combns[combn_count]] = [max(float(x),0) for x in line.split()]
     combn_count += 1
   # Ho Hs Ht Dst Htp Dstp Fst Fstp Fis Fis Dest
   return results
@@ -366,7 +404,7 @@ def parse_fst(parser):
       break
     snp_name,fst = line.split()
     snp_name  = snp_name
-    results[snp_name] = float(fst)
+    results[snp_name] = max(float(fst),0)
   return results
 
 def parse_freq(parser,snp_name):
@@ -399,8 +437,11 @@ def parse_freq(parser,snp_name):
     results[snp_name] = pop_freqs
   return results
 
-def string_parser(content):
-  for line in content:
+# def string_parser(content):
+def process_parser(p):
+  while True:
+    line = p.stdout.readline()
+    if not line: break
     yield line
 
 def tup_eq(a,b):
